Move useWidgets hook to WidgetUtils
